Fix issue with endpoint validation: update SDK to v1.1.2
Original message from Amazon:

Dear Mike,

Thank you for the recent submission of your skill, ‘NYC Transit’.

Your skill submission has failed the certification process.  We’ve included a description of the issue(s) and steps to reproduce below. Please address these and resubmit the skill at your earliest convenience. As a reminder, when you publish a skill, you are eligible for a free Alexa Dev hoodie. Terms and conditions apply.

 Issues with skill endpoint validation

 The skill end-point is not validating the signatures for incoming requests and is accepting requests with an invalid signature URL specified. Please make sure that your signature validation is correct.  To reject an invalid request with an invalid signature or certificate, the skill should respond with HTTP status code 400 (Bad Request) in the response.

 Please refer to the following documentation on how to build your Alexa Skill as a web service and validate requests and signatures.

 https://developer.amazon.com/public/solutions/alexa/alexa-skills-kit/docs/developing-an-alexa-skill-as-a-web-service#verifying-the-signature-certificate-url

Hint from previous email:

We’ve identified that your HTTPS skill endpoint for NYC Transit is not performing full signature verification. This is likely due to a bug in the Alexa Skills Kit Java SDK. We have updated the SpeechletRequestSignatureVerifier class to address the problem. For Java developers using our SDK, we’ve released an update to GitHub and Maven (build version 1.1.2). Please apply this fix to your skill endpoint as soon as possible.

-Alexa Team